# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: UMKM Go API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tickets:
    get:
      operationId: listTickets
      tags: [tickets]
      summary: List support tickets
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [open, in_progress, resolved, closed]
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
      responses:
        "200":
          description: Tickets list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketListResponse"
    post:
      operationId: createTicket
      tags: [tickets]
      summary: Create a support ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicketRequest"
      responses:
        "201":
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}:
    get:
      operationId: getTicket
      tags: [tickets]
      summary: Get ticket detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ticket detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
    put:
      operationId: updateTicket
      tags: [tickets]
      summary: Update ticket (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTicketRequest"
      responses:
        "200":
          description: Updated ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
...
components:
  schemas:
...
    Ticket:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        userEmail:
          type: string
        subject:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        assignedTo:
          type: string
        response:
          type: string
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - userEmail
        - subject
        - description
        - priority
        - status
        - createdAt
    CreateTicketRequest:
      type: object
      properties:
        subject:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, critical]
      required:
        - subject
        - description
        - priority
    UpdateTicketRequest:
      type: object
      properties:
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        assignedTo:
          type: string
        response:
          type: string
        priority:
          type: string
          enum: [low, medium, high, critical]
    TicketListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Ticket"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
...